+++ CHANGELOG.md
@@ -127,3 +127,14 @@
 *   Add `#additions` and `#deletions` to `Rugged::Patch`.
 
     *Mindaugas Mozūras*
+
+
+## Rugged 0.21.2 (November 16, 2014) ##
+
+*   Update bundled libgit2 to 0.21.2 (from 0.21.0).
+
+    See https://github.com/libgit2/libgit2/releases/tag/v0.21.1 and
+    https://github.com/libgit2/libgit2/releases/tag/v0.21.2 for a list
+    of fixed issues.
+
+    *Arthur Schreiber*